Refactor backend structure by renaming container, updating imports, and removing obsolete files; add Dockerignore and test scripts for improved organization and clarity
