+++ deploy/compose/compose.sh
@@ -1,0 +1,1 @@
+docker compose up --build
